pull out sortable_column_controller into separate concern
