+++ app/controllers/concerns/new_inprocess_controller.rb
@@ -1,11 +1,5 @@
 module NewInprocessController
 
-  extend ActiveSupport::Concern
-
-  included do
-    helper_method :sort_column, :sort_direction
-  end
-
   def index
     @order_details = new_or_in_process_orders
                      .includes(
@@ -21,24 +15,6 @@ module NewInprocessController
 
   def new_or_in_process_orders
     raise NotImplementedError
-  end
-
-  def sort_lookup_hash
-    raise NotImplementedError
-  end
-
-  def sort_direction
-    (params[:dir] || "") == "desc" ? "desc" : "asc"
-  end
-
-  def sort_column
-    sort_lookup_hash.key?(params[:sort]) ? params[:sort] : sort_lookup_hash.keys.first
-  end
-
-  def sort_clause
-    Array(sort_lookup_hash[sort_column]).map do |clause|
-      [clause, sort_direction].join(" ")
-    end.join(", ")
   end
 
 end
